added loading & fix my communities box
