+++ dapp/src/components/dashboard/index.jsx
@@ -9,11 +9,10 @@ import { push } from 'connected-react-router'
 import gql from 'graphql-tag'
 import { useQuery } from '@apollo/react-hooks'
 
-// import { getBridgeStatus } from 'selectors/network'
-import { getAccountAddress, getBalances, getProviderInfo } from 'selectors/accounts'
+import { getAccountAddress, getProviderInfo } from 'selectors/accounts'
 import { getHomeNetworkType } from 'selectors/network'
 import { checkIsAdmin } from 'selectors/entities'
-import { getTokenAddressOfByNetwork, getCurrentCommunity } from 'selectors/dashboard'
+import { getCurrentCommunity } from 'selectors/dashboard'
 import { getForeignTokenByCommunityAddress, getHomeTokenByCommunityAddress } from 'selectors/token'
 import { fetchCommunity } from 'actions/token'
 import { loadModal } from 'actions/ui'
@@ -34,7 +33,6 @@ import Users from 'components/dashboard/pages/Users'
 import Businesses from 'components/dashboard/pages/Businesses'
 import JoinBonusPage from 'components/dashboard/pages/JoinBonus'
 import OnRampPage from 'components/dashboard/pages/OnRamp'
-// import SignIn from 'components/common/SignIn'
 
 const GET_COMMUNITY_ORIGIN_NETWORK = (communityAddress) => {
   return gql`
@@ -56,24 +54,22 @@ const DashboardLayout = (props) => {
     networkType,
     accountAddress,
     isAdmin,
-    tokenOfCommunityOnCurrentSide,
     location,
     homeToken,
     fetchEntities,
     push,
     providerInfo,
-    loadModal,
-    dashboard,
-    metadata
+    loadModal
   } = props
   const { address: communityAddress } = useParams()
   const [open, onSetSidebarOpen] = useState(false)
   const [originNetwork, setOriginNetwork] = useState(false)
-  const { fetchCommunityData } = dashboard
   const { loading, error, data } = useQuery(GET_COMMUNITY_ORIGIN_NETWORK(communityAddress))
 
   useEffect(() => {
-    onSetSidebarOpen(false)
+    if (isMobile) {
+      onSetSidebarOpen(false)
+    }
   }, [location.pathname])
 
   useEffect(() => {
@@ -140,14 +136,12 @@ const DashboardLayout = (props) => {
             ? (
               <SidebarContent
                 match={match.url}
-                location={location}
               />
             )
             : <Sidebar
               sidebar={
                 <SidebarContent
                   match={match.url}
-                  location={location}
                 />
               }
               open={open}
@@ -165,18 +159,12 @@ const DashboardLayout = (props) => {
             <div className='content__wrapper'>
               <Switch>
                 {!get(community && community.plugins, 'joinBonus.isRemoved', false) && isAdmin && (
-                  <Route path={`${match.path}/bonus`}
-                    render={() => (
-                      <JoinBonusPage
-                        symbol={foreignToken && foreignToken.symbol}
-                        community={community}
-                        networkType={networkType}
-                        tokenOfCommunityOnCurrentSide={tokenOfCommunityOnCurrentSide}
-                      />
-                    )}
-                  />
+                  <Route exact path={`${match.path}/bonus`}>
+                    <JoinBonusPage />
+                  </Route>
                 )}
-                <Route path={`${match.path}/onramp/:name`}
+
+                <Route exact path={`${match.path}/onramp/:name`}
                   render={({ match }) => (
                     <OnRampPage
                       match={match}
@@ -185,82 +173,40 @@ const DashboardLayout = (props) => {
                     />
                   )}
                 />
+
                 {isAdmin && (foreignToken && foreignToken.tokenType === 'mintableBurnable') && (
-                  <Route
-                    path={`${match.path}/mintBurn`}
-                    render={() => (
-                      <MintBurnPage
-                        symbol={foreignToken && foreignToken.symbol}
-                        tokenAddress={foreignToken.address}
-                        networkType={networkType}
-                        accountAddress={accountAddress}
-                        tokenOfCommunityOnCurrentSide={tokenOfCommunityOnCurrentSide}
-                      />
-                    )}
-                  />
+                  <Route exact path={`${match.path}/mintBurn`}>
+                    <MintBurnPage />
+                  </Route>
                 )}
+
                 {isAdmin && (
-                  <Route
-                    path={`${match.path}/plugins`}
-                    render={() => (
-                      <PluginsPage
-                        community={community}
-                        fetchCommunityData={fetchCommunityData}
-                      />
-                    )}
-                  />
+                  <Route exact path={`${match.path}/plugins`}>
+                    <PluginsPage />
+                  </Route>
                 )}
+
                 {!get((community && community.plugins), 'businessList.isRemoved', false) && (
-                  <Route
-                    exact
-                    path={`${match.path}/merchants`}
-                    render={() => (
-                      <Businesses
-                        // isAdmin={isAdmin}
-                        // community={community}
-                        // networkType={networkType}
-                        // fetchCommunityData={fetchCommunityData}
-                      />
-                    )}
-                  />
+                  <Route exact path={`${match.path}/merchants`}>
+                    <Businesses />
+                  </Route>
                 )}
-                <Route path={`${match.path}/wallet`} render={() => (
-                  <WhiteLabelWallet
-                    value={qrValue}
-                    // fetchCommunityData={fetchCommunityData}
-                  />
-                )} />
-                <Route path={`${match.path}/users`} render={() => (
-                  <Users
-                    // isAdmin={isAdmin}
-                    // community={community}
-                    // networkType={networkType}
-                    // fetchCommunityData={fetchCommunityData}
-                  />
-                )} />
-                <Route
-                  path={`${match.path}/transfer/:sendTo?`}
-                  render={() => (
-                    <TransferPage
-                      loading={fetchCommunityData}
-                      symbol={foreignToken && foreignToken.symbol}
-                      networkType={networkType}
-                      tokenOfCommunityOnCurrentSide={tokenOfCommunityOnCurrentSide}
-                    />
-                  )}
-                />
-                <Route exact path={`${match.path}/:success?`} render={() => (
-                  <Dashboard
-                    loading={fetchCommunityData}
-                    community={community}
-                    foreignToken={foreignToken}
-                    accountAddress={accountAddress}
-                    dashboard={dashboard}
-                    metadata={metadata}
-                    tokenOfCommunityOnCurrentSide={tokenOfCommunityOnCurrentSide}
-                    isAdmin={isAdmin}
-                  />
-                )} />
+
+                <Route exact path={`${match.path}/wallet`}>
+                  <WhiteLabelWallet value={qrValue} />
+                </Route>
+
+                <Route exact path={`${match.path}/users`}>
+                  <Users />
+                </Route>
+
+                <Route exact path={`${match.path}/transfer/:sendTo?`}>
+                  <TransferPage />
+                </Route>
+
+                <Route exact path={`${match.path}/:success?`}>
+                  <Dashboard />
+                </Route>
               </Switch>
             </div>
           </div>
@@ -275,12 +221,8 @@ const mapStateToProps = (state, { match }) => ({
   homeToken: getHomeTokenByCommunityAddress(state, match.params.address),
   foreignToken: getForeignTokenByCommunityAddress(state, match.params.address),
   community: getCurrentCommunity(state, [match.params.address]),
-  metadata: state.entities.metadata,
   isAdmin: checkIsAdmin(state),
-  balances: getBalances(state),
   homeNetwork: getHomeNetworkType(state),
-  dashboard: state.screens.dashboard,
-  tokenOfCommunityOnCurrentSide: getTokenAddressOfByNetwork(state, state.entities.communities[match.params.address]),
   providerInfo: getProviderInfo(state)
 })
