+++ dapp/src/components/home/index.jsx
@@ -1,7 +1,8 @@
 import React, { useState, useEffect } from 'react'
-import Carousel, { Dots } from '@brainhubeu/react-carousel'
+
 import isEmpty from 'lodash/isEmpty'
 import Templates from 'components/home/components/Templates'
+import Tabs from 'components/home/components/Tabs'
 import Faqs from 'components/home/components/Faq'
 import personImage from 'images/person.png'
 import groupImageMobile from 'images/group_mobile.png'
@@ -13,83 +14,75 @@ import withTracker from 'containers/withTracker'
 import { connect } from 'react-redux'
 import { push } from 'connected-react-router'
 import FeaturedCommunities from 'components/home/components/FeaturedCommunities'
-import FeaturedCommunity from 'components/common/FeaturedCommunity'
+// import FeaturedCommunity from 'components/common/FeaturedCommunity'
 import { fetchCommunities } from 'actions/accounts'
 import { getCommunitiesKeys } from 'selectors/accounts'
 
-import Tabs from '@material-ui/core/Tabs'
-import Tab from '@material-ui/core/Tab'
-import Typography from '@material-ui/core/Typography'
-import { makeStyles } from '@material-ui/core/styles'
+// import Tabs from '@material-ui/core/Tabs'
+// import Tab from '@material-ui/core/Tab'
+// import Typography from '@material-ui/core/Typography'
+// import { makeStyles } from '@material-ui/styles'
 
-const toMatrix = (arr, width) =>
-  arr.reduce((rows, key, index) => (index % width === 0 ? rows.push([key])
-    : rows[rows.length - 1].push(key)) && rows, [])
+// const toMatrix = (arr, width) =>
+//   arr.reduce((rows, key, index) => (index % width === 0 ? rows.push([key])
+//     : rows[rows.length - 1].push(key)) && rows, [])
 
-const a11yProps = (index) => ({
-  id: `scrollable-force-tab-${index}`,
-  'aria-controls': `scrollable-force-tabpanel-${index}`
-})
+// const a11yProps = (index) => ({
+//   id: `scrollable-force-tab-${index}`,
+//   'aria-controls': `scrollable-force-tabpanel-${index}`
+// })
 
-const TabPanel = (props) => {
-  const { children, value, index, ...other } = props
+// const TabPanel = (props) => {
+//   const { children, value, index, ...other } = props
 
-  return (
-    <Typography
-      component='div'
-      role='tabpanel'
-      hidden={value !== index}
-      id={`scrollable-force-tabpanel-${index}`}
-      aria-labelledby={`scrollable-force-tab-${index}`}
-      {...other}
-    >
-      <div style={{ padding: '40px 25px' }}>{children}</div>
-    </Typography>
-  )
-}
+//   return (
+//     <Typography
+//       component='div'
+//       role='tabpanel'
+//       hidden={value !== index}
+//       id={`scrollable-force-tabpanel-${index}`}
+//       aria-labelledby={`scrollable-force-tab-${index}`}
+//       {...other}
+//     >
+//       <div style={{ padding: '40px 25px' }}>{children}</div>
+//     </Typography>
+//   )
+// }
 
-const useTabsStyles = makeStyles(theme => ({
-  root: {
-    backgroundColor: '#f8f8f8',
-    alignItems: 'center',
-    minHeight: '80px',
-    borderTopLeftRadius: '10px',
-    borderTopRightRadius: '10px'
-  },
-  indicator: {
-    backgroundColor: ' #052235'
-  }
-}))
+// const useTabsStyles = makeStyles(theme => ({
+//   root: {
+//     backgroundColor: '#f8f8f8',
+//     alignItems: 'center',
+//     minHeight: '80px',
+//     borderTopLeftRadius: '10px',
+//     borderTopRightRadius: '10px'
+//   },
+//   indicator: {
+//     backgroundColor: ' #052235'
+//   }
+// }))
 
-const useTabStyles = makeStyles(theme => ({
-  root: {
-    fontSize: '1.1em',
-    fontFamily: `'Gotham SSm A', 'Gotham SSm B', 'icomoon'`,
-    fontWeight: '500'
-  },
-  textColorPrimary: {
-    color: '#a4a4a4 !important'
-  },
-  selected: {
-    color: '#25435a !important'
-  }
-}))
+// const useTabStyles = makeStyles(theme => ({
+//   root: {
+//     fontSize: '1.1em',
+//     fontFamily: `'Gotham SSm A', 'Gotham SSm B', 'icomoon'`,
+//     fontWeight: '500'
+//   },
+//   textColorPrimary: {
+//     color: '#a4a4a4 !important'
+//   },
+//   selected: {
+//     color: '#25435a !important'
+//   }
+// }))
 
 const HomePage = ({
   accountAddress,
-  communitiesKeys,
-  communities,
-  metadata,
-  fetchCommunities,
   web3connect,
-  push
+  fetchCommunities,
+  push,
+  communitiesKeys
 }) => {
-  const [value, setValue] = useState(0)
-  const [valueSpinner, onChangeSpinner] = useState(0)
-
-  const tabsClasses = useTabsStyles()
-  const tabClasses = useTabStyles()
-
   useEffect(() => {
     if (accountAddress) {
       fetchCommunities(accountAddress)
@@ -109,12 +102,6 @@ const HomePage = ({
     }
   }
 
-  const communitiesIOwn = React.useMemo(() => {
-    return communitiesKeys
-      .map((communityAddress) => communities[communityAddress])
-      .filter(obj => !!obj).filter(({ isAdmin, token }) => isAdmin && token)
-  }, [communitiesKeys, communities])
-
   const showDashboard = (communityAddress, name) => {
     if (window && window.analytics) {
       if (name) {
@@ -123,38 +110,6 @@ const HomePage = ({
     }
     push(`/view/community/${communityAddress}`)
   }
-
-  const slides = React.useMemo(() => {
-    if (!isEmpty(communitiesIOwn)) {
-      const myCommunities = communitiesIOwn.map((entity, index) => {
-        const { community, token } = entity
-        const { communityAddress } = community
-        return (
-          <div className='cell medium-12 small-24' key={community && community.name}>
-            <FeaturedCommunity
-              accountAddress={accountAddress}
-              symbol={token && token.symbol}
-              metadata={{
-                ...metadata[token.tokenURI],
-                ...metadata[community.communityURI]
-              }}
-              tokenURI={token && token.tokenURI}
-              showDashboard={() => showDashboard(communityAddress)}
-              community={community}
-            />
-          </div>
-        )
-      })
-      const myCommunitiesList = toMatrix(myCommunities, 4).map((items, index) => (
-        <div style={{ width: '100%' }} key={index} className='grid-x grid-margin-x grid-margin-y'>
-          {items}
-        </div>
-      ))
-      return myCommunitiesList
-    }
-  }, [communitiesIOwn])
-
-  const handleChange = (event, newValue) => setValue(newValue)
 
   return (
     <div className='home_page'>
@@ -188,48 +143,10 @@ const HomePage = ({
           <div className='grid-x align-justify grid-margin-x grid-margin-y'>
             <div className='cell medium-24 large-12'>
               {
-                accountAddress && !isEmpty(communitiesIOwn) ? (
-                  <div className='home_page__tabs'>
-                    <Tabs
-                      classes={tabsClasses}
-                      value={value}
-                      onChange={handleChange}
-                      variant='scrollable'
-                      scrollButtons='on'
-                      indicatorColor='primary'
-                      textColor='inherit'
-                    >
-                      <Tab disabled={isEmpty(communitiesIOwn)} label='My communities' classes={tabClasses} {...a11yProps(0)} />
-                      <Tab label='Templates' classes={tabClasses} {...a11yProps(1)} />
-                    </Tabs>
-                    <TabPanel value={value} index={0}>
-                      <Carousel
-                        value={valueSpinner}
-                        centered
-                        infinite
-                        draggable
-                        onChange={onChangeSpinner}
-                        animationSpeed={1000}
-                        slidesPerPage={1}
-                        breakpoints={{
-                          1000: {
-                            slidesPerPage: 2
-                          },
-                          800: {
-                            slidesPerPage: 1
-                          }
-                        }}
-                      >
-                        {slides}
-                      </Carousel>
-                      <Dots value={valueSpinner} onChange={onChangeSpinner} number={slides && slides.length} />
-                    </TabPanel>
-                    <TabPanel value={value} index={1}>
-                      <Templates showIssuance={showIssuance} />
-                    </TabPanel>
-                  </div>
+                !isEmpty(communitiesKeys) ? (
+                  <Tabs showDashboard={showDashboard} showIssuance={showIssuance} />
                 ) : (
-                  <Templates withDecoration showIssuance={showIssuance} />
+                  <Templates withDecoration showDashboard={showDashboard} showIssuance={showIssuance} />
                 )
               }
             </div>
@@ -246,8 +163,6 @@ const HomePage = ({
 
 const mapStateToProps = (state) => ({
   accountAddress: state.network.accountAddress,
-  metadata: state.entities.metadata,
-  communities: state.entities.communities,
   communitiesKeys: getCommunitiesKeys(state)
 })
