+++ dapp/src/components/oven/CommunitiesList.jsx
@@ -77,12 +77,7 @@ export default ({
                     return (
                       <div className='medium-12 large-8 small-24 cell' key={address}>
                         <FeaturedCommunity
-                          accountAddress={account}
-                          metadata={{
-                            ...metadata[token.tokenURI],
-                            ...metadata[community.communityURI]
-                          }}
-                          tokenURI={token && token.tokenURI}
+                          token={token}
                           showDashboard={() => showDashboard(address, community.name)}
                           community={community}
                         />
