+++ dapp/src/components/wizard/pages/DetailsStep.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { connect, getIn } from 'formik'
 import filter from 'lodash/filter'
-import { withStyles } from '@material-ui/core/styles'
+import { withStyles } from '@material-ui/styles'
 import MuiExpansionPanel from '@material-ui/core/ExpansionPanel'
 import MuiExpansionPanelSummary from '@material-ui/core/ExpansionPanelSummary'
 import MuiExpansionPanelDetails from '@material-ui/core/ExpansionPanelDetails'
