+++ dapp/src/constants/uiConstants.js
@@ -1,12 +1,5 @@
 export const WRONG_NETWORK_MODAL = 'WRONG_NETWORK_MODAL'
-export const EXCHANGE_MODAL = 'EXCHANGE_MODAL'
-export const LOADING_MODAL = 'LOADING_MODAL'
-export const BUSINESS_LIST_MODAL = 'BUSINESS_LIST_MODAL'
-export const BRIDGE_MODAL = 'BRIDGE_MODAL'
-export const SIMPLE_EXCHANGE_MODAL = 'SIMPLE_EXCHANGE_MODAL'
-export const ECONOMIC_CALCULATOR_MODAL = 'ECONOMIC_CALCULATOR_MODAL'
 export const ADD_BUSINESS_MODAL = 'ADD_BUSINESS_MODAL'
-export const NO_DATA_ABOUT_OWNER_MODAL = 'NO_DATA_ABOUT_OWNER_MODAL'
 export const SHOW_MORE_MODAL = 'SHOW_MORE_MODAL'
 export const QR_MODAL = 'QR_MODAL'
 export const ADD_USER_MODAL = 'ADD_USER_MODAL'
