+++ dapp/src/reducers/screens/dashboard.js
@@ -5,9 +5,9 @@ import pick from 'lodash/pick'
 export default (state = {}, action) => {
   switch (action.type) {
     case FETCH_COMMUNITY_DATA.SUCCESS:
-      return { ...state, fetchCommunityData: false, communityAddress: action.communityAddress }
+      return { ...state, communityAddress: action.communityAddress }
     case FETCH_COMMUNITY_DATA.REQUEST:
-      return { ...state, fetchCommunityData: true }
+      return { ...state, communityAddress: '', totalSupply: '' }
     case FETCH_TOKEN_TOTAL_SUPPLY.SUCCESS:
       return { ...state, totalSupply: { ...state.totalSupply, [action.tokenAddress]: action.response.totalSupply } }
     case FETCH_TOKEN_PROGRESS.SUCCESS:
